have ui elements functions need data from backend
